-added date conversion to database class -added date insert into db method -added test into dbtest
